requirements.txt
django
psycopg2
djangorestframework
markdown
django-filter
djangorestframework-simplejwt
inject
factory-boy
Faker
black